feat(create gif): employees can post gifs
- create gif component
- create form data apiRequest helper

[Finishes #card-29908980]
