+++ package.json
@@ -7,6 +7,7 @@
     "@material-ui/icons": "^4.5.1",
     "@material-ui/styles": "^4.6.0",
     "@tinymce/tinymce-react": "^3.3.2",
+    "material-ui-dropzone": "^2.4.7",
     "moment": "^2.24.0",
     "notistack": "^0.9.6",
     "react": "^16.12.0",
